Folder Types Component Chnages

+++ out/extension.js
@@ -77,7 +77,7 @@ function activate(context) {
                 case 'loadComponents':
                     if (message.type) {
                         var sourceOrg = orgsList.find((org) => org.orgId === message.sourceOrgId);
-                        getComponents(sourceOrg.accessToken, sourceOrg.instanceUrl, message.type)
+                        getComponents(sourceOrg.accessToken, sourceOrg.instanceUrl, message.type, message.isFolder)
                             .then((data) => {
                             panel.webview.postMessage({ command: 'components', components: data, type: message.type });
                         });
@@ -174,7 +174,7 @@ function activate(context) {
                                 if (result.done === 'true') {
                                     clearInterval(intervalId);
                                     sourceOrgFiles = result.fileNames;
-                                    extractComponents(result.zipFile, tmpDirectory + "/" + time, sourceOrg.alias);
+                                    extractComponents(result.zipFile, tmpDirectory, '' + time, sourceOrg.alias);
                                     if (destOrgFiles.size > 0) {
                                         postCompareResults(sourceOrgFiles, destOrgFiles, tmpDirectory + "/" + time + "/" + sourceOrg.alias, tmpDirectory + "/" + time + "/" + destOrg.alias, panel);
                                     }
@@ -192,7 +192,7 @@ function activate(context) {
                                 if (result.done === 'true') {
                                     clearInterval(intervalId);
                                     destOrgFiles = result.fileNames;
-                                    extractComponents(result.zipFile, tmpDirectory + "/" + time, destOrg.alias);
+                                    extractComponents(result.zipFile, tmpDirectory, '' + time, destOrg.alias);
                                     if (sourceOrgFiles.size > 0) {
                                         postCompareResults(sourceOrgFiles, destOrgFiles, tmpDirectory + "/" + time + "/" + sourceOrg.alias, tmpDirectory + "/" + time + "/" + destOrg.alias, panel);
                                     }
@@ -226,18 +226,21 @@ function postCompareResults(sourceOrgFiles, destOrgFiles, sourceOrgPath, destOrg
     });
     panel.webview.postMessage({ command: 'compareResults', files: files });
 }
-function extractComponents(zipfile, tmpPath, directory) {
+function extractComponents(zipfile, tmpPath, subDir, alias) {
     const buffer = Buffer.from(zipfile, 'base64');
     if (!fs.existsSync(tmpPath)) {
         fs.mkdirSync(tmpPath);
     }
-    const zipFilePath = path.join(tmpPath, directory + '.zip');
+    if (!fs.existsSync(tmpPath + "/" + subDir)) {
+        fs.mkdirSync(tmpPath + "/" + subDir);
+    }
+    const zipFilePath = path.join(tmpPath + "/" + subDir, alias + '.zip');
     fs.writeFileSync(zipFilePath, buffer);
-    if (!fs.existsSync(tmpPath + "/" + directory)) {
-        fs.mkdirSync(tmpPath + "/" + directory);
+    if (!fs.existsSync(tmpPath + "/" + subDir + "/" + alias)) {
+        fs.mkdirSync(tmpPath + "/" + subDir + "/" + alias);
     }
     const zip = new AdmZip(zipFilePath);
-    zip.extractAllTo(tmpPath + "/" + directory, true);
+    zip.extractAllTo(tmpPath + "/" + subDir + "/" + alias, true);
 }
 function cancelDeploy(accessToken, endPoint, deployJobId) {
     return new Promise((resolve, reject) => {
@@ -328,40 +331,60 @@ function retrieve(accessToken, endPoint, packagexml) {
         });
     });
 }
-function getComponents(accessToken, endPoint, type) {
+function getComponents(accessToken, endPoint, type, isFolder) {
     return new Promise((resolve, reject) => {
-        sendSoapReuest(accessToken, endPoint, '<met:listMetadata><met:queries><met:type>' + type + '</met:type></met:queries></met:listMetadata>')
+        sendSoapReuest(accessToken, endPoint, '<met:listMetadata><met:queries><met:type>' + type + (isFolder ? 'Folder' : '') + '</met:type></met:queries></met:listMetadata>')
             .then((result) => {
             const comps = result['soapenv:Envelope']['soapenv:Body']['listMetadataResponse'];
-            if (comps !== "") {
-                if (comps['result'] instanceof Array) {
-                    const compsList = comps['result'].map((comp) => ({
-                        name: comp['fullName'],
-                        type: comp['type'],
-                        lastModifiedByName: comp['lastModifiedByName'],
-                        lastModifiedDate: new Date(comp['lastModifiedDate']).toLocaleDateString(),
-                        manageableState: comp['manageableState']
-                    }));
-                    resolve(compsList);
-                }
-                else {
-                    resolve([{
-                            name: comps['result']['fullName'],
-                            type: comps['result']['type'],
-                            lastModifiedByName: comps['result']['lastModifiedByName'],
-                            lastModifiedDate: new Date(comps['lastModifiedDate']).toLocaleDateString(),
-                            manageableState: comps['manageableState']
-                        }]);
-                }
+            let results = buildComponents(comps);
+            if (isFolder) {
+                let folderresults = [];
+                const promises = results.map((element) => {
+                    return sendSoapReuest(accessToken, endPoint, '<met:listMetadata><met:queries><met:type>' + type +
+                        '</met:type><met:folder>' + element.name + '</met:folder></met:queries></met:listMetadata>')
+                        .then((result) => {
+                        const comps = result['soapenv:Envelope']['soapenv:Body']['listMetadataResponse'];
+                        let fldresults = buildComponents(comps);
+                        folderresults = [...folderresults, ...fldresults];
+                    });
+                });
+                Promise.all(promises)
+                    .then(() => {
+                    resolve(folderresults);
+                });
             }
             else {
-                resolve([]);
+                resolve(results);
             }
         })
             .catch((error) => {
             reject(error);
         });
     });
+}
+function buildComponents(comps) {
+    let results = [];
+    if (comps !== "") {
+        if (comps['result'] instanceof Array) {
+            results = comps['result'].map((comp) => ({
+                name: comp['fullName'],
+                type: comp['type'],
+                lastModifiedByName: comp['lastModifiedByName'],
+                lastModifiedDate: new Date(comp['lastModifiedDate']).toLocaleDateString(),
+                manageableState: comp['manageableState']
+            }));
+        }
+        else {
+            results.push({
+                name: comps['result']['fullName'],
+                type: comps['result']['type'],
+                lastModifiedByName: comps['result']['lastModifiedByName'],
+                lastModifiedDate: new Date(comps['lastModifiedDate']).toLocaleDateString(),
+                manageableState: comps['manageableState']
+            });
+        }
+    }
+    return results;
 }
 function getTypes(accessToken, endPoint, favorites) {
     return new Promise((resolve, reject) => {
@@ -373,7 +396,8 @@ function getTypes(accessToken, endPoint, favorites) {
                 typesList.push({
                     name: element['xmlName'],
                     isFavorite: favorites.indexOf(element['xmlName']) >= 0,
-                    hidden: false
+                    hidden: false,
+                    inFolder: element['inFolder']
                 });
                 if (element['childXmlNames']) {
                     if (element['childXmlNames'] instanceof Array) {
@@ -381,7 +405,8 @@ function getTypes(accessToken, endPoint, favorites) {
                             typesList.push({
                                 name: childname,
                                 isFavorite: favorites.indexOf(element['xmlName']) >= 0,
-                                hidden: false
+                                hidden: false,
+                                inFolder: 'false'
                             });
                         });
                     }
@@ -389,7 +414,8 @@ function getTypes(accessToken, endPoint, favorites) {
                         typesList.push({
                             name: element['childXmlNames'],
                             isFavorite: favorites.indexOf(element['xmlName']) >= 0,
-                            hidden: false
+                            hidden: false,
+                            inFolder: 'false'
                         });
                     }
                 }
@@ -669,11 +695,12 @@ function getWebviewContent(basedpath, scriptUri, cssUri) {
 			</html>`;
 }
 function deactivate() {
-    if (tmpDirectory) {
-        /*try {
+    if (tmpDirectory && fs.existsSync(tmpDirectory)) {
+        try {
             fs.rmSync(tmpDirectory, { recursive: true, force: true });
-        } catch (err) {
-        }*/
+        }
+        catch (err) {
+        }
     }
 }
 //# sourceMappingURL=extension.js.map
